src/cogs/role_cog.py
"""
Role Management Cog

This cog handles Discord role management and Minecraft role synchronization.
Can be disabled by not loading this cog.
"""

import discord
from discord.ext import commands
from discord import app_commands
import os
import logging
from typing import Optional

logger = logging.getLogger(__name__)

class RoleCog(commands.Cog):
    """Cog for handling role management."""
    
    def __init__(self, bot: commands.Bot):
        self.bot = bot
        self.whitelist_channel_id = int(os.getenv("WHITELIST_CHANNEL_ID", "0"))
        
        # Check if role management is enabled
        if not self.whitelist_channel_id:
            logger.warning("Role cog disabled - WHITELIST_CHANNEL_ID not configured")
            return
        
        logger.info("Role management cog initialized")
    
    @app_commands.command(name="role", description="Manage your Discord roles")
    async def role_command(self, interaction: discord.Interaction):
        """Handle role management requests."""
        if not self.whitelist_channel_id:
            await interaction.response.send_message(
                "Role management functionality is currently disabled.",
                ephemeral=True
            )
            return
        
        # This would contain the role management logic
        await interaction.response.send_message(
            "Role management functionality is available but not fully implemented in this cog.",
            ephemeral=True
        )

async def setup(bot: commands.Bot):
    """Setup function for the cog."""
    await bot.add_cog(RoleCog(bot)) 